Global Type and Color Tweaks
Tweaking link colors to be on brand and match the blog.  Had some
trouble making all links Growth-3 because all a's are defined in one
place and inherit their color.  Eventually we'll want to define the a's
on white backgrounds to be one color and a's on another background to
be a seperate color (if only for contrast).